Convert api to services

+++ Shell/Views/Frm/frmLogin.cs
@@ -3,8 +3,12 @@ using Google.Apis.Discovery.v1;
 using Google.Apis.Discovery.v1.Data;
 using Google.Apis.Services;
 using Ji;
+using Ji.Commons;
 using Ji.Core;
 using Ji.Model;
+using Ji.Model.CustomModels;
+using Ji.Model.LoginModels;
+using Ji.Services.Interface;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using Shell.Model;
@@ -22,69 +26,11 @@ namespace Shell.Views.Frm
 {
     public partial class frmLogin : ClientForm
     {
-        protected string googleplus_client_id = "530840873560-39mgdeqjhf2ns2vueodsrg011sdj781p.apps.googleusercontent.com";    // Replace this with your Client ID
-        protected string googleplus_client_secret = "F2YHsmQiqd97jeHCLk-RTyva";                                                // Replace this with your Client Secret
-        protected string googleplus_redirect_url = "http://localhost:2443/Index.aspx";                                         // Replace this with your Redirect URL; Your Redirect URL from your developer.google application should match this URL.
-        protected string Parameters;
+        private readonly ILoginServices _loginServices;
         public frmLogin()
         {
+            _loginServices = _loginServices.GetServices();
             InitializeComponent();
-            //frmLoginWithGoogle frm = new frmLoginWithGoogle();
-            //frm.Show();
-        }
-        private async Task Run()
-        {
-            // Create the service.
-            var service = new DiscoveryService(new BaseClientService.Initializer
-            {
-                ApplicationName = "Discovery Sample",
-                ApiKey = "[YOUR_API_KEY_HERE]",
-            });
-
-            // Run the request.
-            Console.WriteLine("Executing a list request...");
-            var result = await service.Apis.List().ExecuteAsync();
-
-            // Display the results.
-            if (result.Items != null)
-            {
-                foreach (DirectoryList.ItemsData api in result.Items)
-                {
-                    Console.WriteLine(api.Id + " - " + api.Title);
-                }
-            }
-        }
-        private void btnGoogleLogin()
-        {
-            HttpWebRequest webRequest = (HttpWebRequest)WebRequest.Create("https://accounts.google.com/o/oauth2/token");
-            webRequest.Method = "POST";
-            Parameters = "client_id=" + googleplus_client_id + "&client_secret=" + googleplus_client_secret + "&grant_type=authorization_code";
-            byte[] byteArray = Encoding.UTF8.GetBytes(Parameters);
-            webRequest.ContentType = "application/x-www-form-urlencoded";
-            webRequest.ContentLength = byteArray.Length;
-            Stream postStream = webRequest.GetRequestStream();
-            // Add the post data to the web request
-            postStream.Write(byteArray, 0, byteArray.Length);
-            postStream.Close();
-
-            WebResponse response = webRequest.GetResponse();
-            postStream = response.GetResponseStream();
-            StreamReader reader = new StreamReader(postStream);
-            string responseFromServer = reader.ReadToEnd();
-
-            GooglePlusAccessToken serStatus = JsonConvert.DeserializeObject<GooglePlusAccessToken>(responseFromServer);
-
-            if (serStatus != null)
-            {
-                string accessToken = string.Empty;
-                accessToken = serStatus.access_token;
-
-                if (!string.IsNullOrEmpty(accessToken))
-                {
-                    // This is where you want to add the code if login is successful.
-                    // getgoogleplususerdataSer(accessToken);
-                }
-            }
         }
         /// <summary>
         /// Hàm đăng nhập khi bấm nút btnLogin
@@ -102,40 +48,43 @@ namespace Shell.Views.Frm
             }
             else
             {
-                string Username = txtUsername.Text.Trim().ToLower();
-                string Password = txtPassword.Text.Trim();
-                //L_Users user = db.L_Users.FirstOrDefault(x => x.Username.Equals(Username) && x.Password.Equals(Password));
-                string user = API.API_LOGIN(Extension.GetAppSetting("Token"), Username, Password);
-                SplashScreenManager.CloseForm();
-                JObject json = JObject.Parse(user);
-                if (json["error_description"] != null)
+                UI.ShowSplashForm();
+                LoginRequest login = new LoginRequest();
+                login.Username = txtUsername.Text;
+                login.Password = txtPassword.Text;
+                ResultCustomModel<LoginResultModel> result = _loginServices.UserLogin(login);
+                UI.CloseSplashForm();
+                if (result.Success)
                 {
-                    UI.Warning(json["error_description"].ToString());
-                    return;
-                }
-                Extension.Setup = json;
-                API.Access_Token = json["token"]?.ToString() ?? "";
-                SplashScreenManager.ShowForm(typeof(PleaseWaiting));
-                if (chkRememberPassword.Checked)
-                {
-                    if (!string.IsNullOrEmpty(Username) && !string.IsNullOrEmpty(Password))
+                    AuthorizeConstant.Users = result.Data.User;
+                    AuthorizeConstant.Token = result.Data.Token;
+
+                    if (chkRememberPassword.Checked)
+                    {
                         new XDocument(
                             new XElement("root",
-                                new XElement("User", Username),
-                                new XElement("Password", txtPassword.Text.Trim())
+                                new XElement("User", login.Username),
+                                new XElement("Password", login.Password)
                             )
                         )
                         .Save("Information.xml");
+                    }
+                    //Nếu không lựa chọn Remember Password thì xoá luôn file lưu trữ password
+                    else
+                    {
+                        if (File.Exists("Information.xml"))
+                        {
+                            File.Delete("Information.xml");
+                        }
+                    }
+                    DialogResult = DialogResult.OK;
+                    Close();
                 }
                 else
                 {
-                    if (File.Exists("Information.xml"))
-                    {
-                        File.Delete("Information.xml");
-                    }
+                    UI.ShowError(result.Message["vn"]);
+                    return;
                 }
-                this.DialogResult = DialogResult.OK;
-                Close();
             }
         }
 
@@ -151,50 +100,38 @@ namespace Shell.Views.Frm
         {
             if (File.Exists("Information.xml"))
             {
-                if (!string.IsNullOrEmpty(GetInfoByXML("User")) && !string.IsNullOrEmpty(GetInfoByXML("Password")))
+                if (!string.IsNullOrEmpty(Extension.GetInfoByXML("Information.xml","User")) && !string.IsNullOrEmpty(Extension.GetInfoByXML("Information.xml", "Password")))
                 {
-                    string Username = GetInfoByXML("User"), Password = GetInfoByXML("Password");
-                    string user = API.API_LOGIN(Extension.GetAppSetting("Token"), Username, Password);
-                    JObject json = JObject.Parse(user);
-                    if (json["error_description"] != null)
+                    UI.ShowSplashForm();
+                    LoginRequest login = new LoginRequest();
+                    login.Username = Extension.GetInfoByXML("Information.xml", "User");
+                    login.Password = Extension.GetInfoByXML("Information.xml", "Password");
+                    ResultCustomModel<LoginResultModel> result = _loginServices.UserLogin(login);
+                    UI.CloseSplashForm();
+                    if (result.Success)
                     {
-                        UI.Warning(json["error_description"].ToString());
-                        return;
+                        AuthorizeConstant.Users = result.Data.User;
+                        AuthorizeConstant.Token = result.Data.Token;
+                        DialogResult = DialogResult.OK;
+                        Close();
                     }
                     else
                     {
-                        Extension.Setup = json;
-                        API.Access_Token =  json["token"].ToString();
-                        DialogResult = DialogResult.OK;
-                        Close();
+                        UI.ShowError(result.Message["vn"]);
+                        return;
                     }
                 }
                 else
                 {
                     if (flag)
                     {
-                        if (SplashScreenManager.Default != null && SplashScreenManager.Default.IsSplashFormVisible)
-                            SplashScreenManager.CloseForm();
+                        UI.CloseSplashForm();
                         File.Delete("Information.xml");
-                        UI.Debug("Tài khoản và mật khẩu của bạn đã lưu không chính xác, vui lòng bấm đăng nhập lần nữa để thử lại!");
+                        UI.Debug(MessageConstant.InvalidInfoSave);
                     }
                 }
-            }           
-            if (SplashScreenManager.Default != null && SplashScreenManager.Default.IsSplashFormVisible)
-                SplashScreenManager.CloseForm();
+            }
 
-        }
-        /// <summary>
-        /// Lấy thông tin theo element trong file XML
-        /// </summary>
-        /// <param name="el"></param>
-        /// <returns></returns>
-        private string GetInfoByXML(string el)
-        {
-            XElement root = XElement.Load("Information.xml");
-            if (root.Element(el) != null)
-                return root.Element(el).Value;
-            return null;
         }
 
         private void btnExit_Click(object sender, EventArgs e)
